Fix another x64 drag-n-drop bug
git-svn-id: https://svn.isis.vanderbilt.edu/GMESRC/trunk@1831 b35830b6-564f-0410-ac5f-b387e50cb686

+++ GME/Gme/GMEOLEData.cpp
@@ -71,7 +71,7 @@ void CGMEDataDescriptor::Serialize(CArchive& ar)
 		while(pos)
 			ar << *(rects.GetNext(pos));
 		
-		ar << annRects.GetCount();
+		ar << (int)annRects.GetCount();
 		pos = annRects.GetHeadPosition();
 		while(pos)
 			ar << *(annRects.GetNext(pos));
